@@ -36,12 +36,12 @@ namespace AllInOne_Learning.ExtensionsForConfigureServices
             .AddCookie(options =>
             {
                 options.LoginPath = "/con/googleapi/index"; //this is basically reroutes whenever u try to access unauthorized page
-            })
-            .AddGoogle(options =>
-            {
-                options.ClientId = "741556789535-s3tufo74m2arffip4b4rfaopt8b8eetb.apps.googleusercontent.com";
-                options.ClientSecret = "ZTcOlxyxtTJCpI74Ox2FuRWt";
             });
+            //.AddGoogle(options =>
+            //{
+            //    options.ClientId = "741556789535-s3tufo74m2arffip4b4rfaopt8b8eetb.apps.googleusercontent.com";
+            //    options.ClientSecret = "ZTcOlxyxtTJCpI74Ox2FuRWt";
+            //});
             #endregion
             return services;
         }
