@@ -8,7 +8,10 @@ import SqlInterview from './Components/SqlInterview/SqlInterview';
 import Pending from './Components/PendingImplementation/Pending';
 import ReactBasics from './Components/ReactBasics/ReactBasics';
 import ReactInterview from './Components/ReactInterview/ReactInterview';
+import Azure from './Components/Azure/Azure';
 import ReactImplementation from './ReactImplementations/ReactImplementations';
+import Docker from './Components/Docker/Docker';
+import OtherBasicImplementation from './Components/OtherBasicImplementation/OtherBasicImplementation';
 
 class App extends Component {
   state={
@@ -18,8 +21,11 @@ class App extends Component {
       {Name:'DotNet Interview',Path:'/interview'},
       {Name:'Sql Interview',Path:'/sqlinterview'},
       {Name:'React Interview',Path:'/reactinterview'},
+      {Name:'Azure',Path:'/azure'},
       {Name:'React Implementations',Path:'/reactimplementation'},
-      {Name:'Pending Implementation',Path:'/pending'}
+      {Name:'Docker',Path:'/docker'},
+      {Name:'Pending Implementation',Path:'/pending'},
+      {Name:'Other Implementations',Path:'/otherbasicimplementation'},
     ]
   }
   render() {
@@ -33,7 +39,10 @@ class App extends Component {
           <Route path='/sqlinterview' component={SqlInterview} />
           <Route path='/pending' component={Pending} />
           <Route path='/reactinterview' component={ReactInterview} />
+          <Route path='/azure' component={Azure} />
+          <Route path='/docker' component={Docker} />
           <Route path='/reactimplementation' component={ReactImplementation} />
+          <Route path='/otherbasicimplementation' component={OtherBasicImplementation} />
         </Switch>
       </div>
     );
